Rename validation_schemas.py to response_schemas.py
